Add files via upload
used a simple yet efficient technique where i simply traversed the array and checked whether the number is odd or not and maintained a streak which when reached three would return true else the streak would go back to 0 and traversing will continue. Outside this loop we return false since we were unsuccessful to find the trio of consecutive odd numbers.